pyproject.toml
[project]
name = "entan-mkdocs"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.13"
dependencies = [
    "httpx>=0.28.1",
    "markdown2>=2.5.4",
    "pandas>=2.3.3",
    "pytest>=9.0.2",
    "pytest-asyncio>=1.3.0",
    "pytest-cov>=7.0.0",
    "pytest-mock>=3.15.1",
    "python-fasthtml>=0.12.35",
    "passlib>=1.7.4",
    "resend>=2.19.0",
    "tabulate>=0.9.0",
    "uvicorn<=0.38.0",
    "bcrypt==4.0.1"
]
